backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
groq
firebase-admin
python-telegram-bot
httpx
sqlalchemy
asyncpg
PyMuPDF
ollama
boto3
python-multipart
greenlet
psycopg2-binary
pgvector
pyyaml
openai
pdfplumber
edge-tts